Improved version
now it contains an actual header, and code section and stuff. There was an attempt of making routines and assembly code for operations. But there is something wrong (6 + 2 is giving 29 lol). Maybe something is wrong in the stack? Or maybe wrong registers? Or even just bad conversion sending it to serial monitor. Anyways, its a start to improve on.